requirements.txt
pandas~=2.2.3
numpy
djangorestframework
psycopg2
psycopg
django~=5.1.2
requests~=2.32.3
